+++ src/main/java/com/hotdeal/discord/infrastructure/discord/config/DiscordBotInitializer.java
@@ -1,6 +1,6 @@
 package com.hotdeal.discord.infrastructure.discord.config;
 
-import com.hotdeal.discord.infrastructure.discord.command.CommandListener;
+import com.hotdeal.discord.infrastructure.discord.listener.CommandListener;
 import com.hotdeal.discord.common.exception.ErrorCode;
 import com.hotdeal.discord.infrastructure.discord.exception.DiscordBotInitializationException;
 import jakarta.annotation.PostConstruct;
